requirements.txt
mss
kafka-python
inference-sdk
ultralytics
opencv-python
Flask
python-dotenv
deep-sort-realtime